Major refactor with Malachi
Reviewed layout and refactored into a more object/MVC oriented model.
