--- src/main/java/com/turn_based_game/CLI_view/BattleMenu.java
@@ -1,10 +1,0 @@
-package com.turn_based_game.CLI_view;
-
-public class BattleMenu{
-
-    
-
-    public void Open() {
-
-    }
-}
