--- src/main/java/com/turn_based_game/Target.java
@@ -1,6 +1,0 @@
-package com.turn_based_game;
-
-public enum Target {
-    self,
-    enemy
-}
