+++ src/main/java/com/turn_based_game/controller/BattleController.java
@@ -1,0 +1,18 @@
+package com.turn_based_game.controller;
+
+import com.turn_based_game.model.Battle;
+import com.turn_based_game.model.Combatant;
+import com.turn_based_game.model.actions.Action;
+
+import java.io.InputStream;
+import java.io.OutputStream;
+
+public class BattleController {
+       // Instance variables
+
+
+        // Constructors
+
+        // Behaviors
+
+    }
