+++ src/main/java/com/turn_based_game/model/Enemy.java
@@ -1,12 +1,13 @@
-package com.turn_based_game;
+package com.turn_based_game.model;
 
-import com.turn_based_game.actions.Action;
+import com.turn_based_game.model.actions.Action;
 
+@Deprecated
 public class Enemy extends Combatant {
 
     // Constructors
     public Enemy(String name, int maxHealth, int maxDefence, Action[] actionsAvailable) {
-        super(name, maxHealth, maxDefence, actionsAvailable);
+        super(name, maxHealth, maxDefence, actionsAvailable, false);
     }
 
     // Behaviors
@@ -14,6 +15,7 @@ public class Enemy extends Combatant {
     //Choose a random action from actions
     // TODO: implement this.
     @Override
+    @Deprecated
     public Action promptForAction(){
         return null;
     }
