+++ src/main/java/com/turn_based_game/view/web/Menu.java
@@ -1,0 +1,6 @@
+package com.turn_based_game.view.web;
+
+//this is a placeholder for structure's sake.
+
+public class Menu {
+}
